works acceptably with 6.0. if you run into trouble, change the C language standard. also, you get an error before your sensor is detected. this sorts its motherfucking self
Former-commit-id: 2941f7b6ac69c10054fdb93c3ea2e10a619c032f

+++ metaioSDK.framework/Headers/MetaioCloudPlugin/MetaioCloudPluginDelegate.h
@@ -1,17 +1,11 @@
-//
-//  MetaioCloudPluginDelegate.h
-//  MetaioCloudPlugin
-//
-//  Created by Stefan Misslinger.
-//  Copyright 2013 metaio, Inc. All rights reserved.
-//
+// Copyright 2007-2014 Metaio GmbH. All rights reserved.
 
 @protocol MetaioCloudPluginDelegate
 
 /** 
  * Provide the channel number that should be opened by the plugin 
  */
-- (NSInteger) getChannelID;
+- (int) getChannelID;
 
 
 @optional
